[test] pull out test functor
